WIP
Reviewed by NOBODY (OOPS!).

+++ LayoutTests/http/tests/navigation/resources/check-ping.py
@@ -7,21 +7,28 @@ from ping_file_path import ping_filepath
 
 sys.stdout.write('Content-Type: text/html\r\n\r\n')
 
-while not os.path.isfile(ping_filepath):
+iterations = 100
+while not os.path.isfile(ping_filepath) and iterations != 0:
     time.sleep(0.01)
+    iterations -= 1
 
-sys.stdout.write(
-    '<html><body>\n'
-    'Ping sent successfully'
-)
+if iterations == 0:
+    sys.stdout.write(
+        '<html><body>\n'
+        'Ping not sent'
+    )
+else:
+    sys.stdout.write(
+        '<html><body>\n'
+        'Ping sent successfully'
+    )
 
-ping_file = open(ping_filepath, 'r')
-for line in ping_file.readlines():
-    sys.stdout.write('<br>{}'.format(line.strip()))
+    with open(ping_filepath, 'r') as ping_file:
+        for line in ping_file.readlines():
+            sys.stdout.write('<br>{}'.format(line.strip()))
 
-ping_file.close()
-if os.path.isfile(ping_filepath):
-    os.remove(ping_filepath)
+    if os.path.isfile(ping_filepath):
+        os.remove(ping_filepath)
 
 sys.stdout.write(
     '<script>'
